Fixed permissions in sort and enabled to work with python 3 (as well as 2).
Former-commit-id: f0b3deb66c7c5401f5a9a58570b0e88b263ad155

+++ python/examples/sort/src/sort.py
@@ -114,7 +114,7 @@ def sort(nums_file, fragments, num_range):
         num_range = int(len(nums) / fragments)
         print("Using num_range: %s" % num_range)
 
-    nums_per_node = len(nums) / fragments
+    nums_per_node = int(len(nums) / fragments)
     partial_result = {}
     result = []
